Update CoverAlreadyExistsSpecification.cs
* Reorder checks to exit early before GdiPlus check, this makes it possible to test assertion of different file sizes

+++ src/NzbDrone.Core/MediaCover/CoverAlreadyExistsSpecification.cs
@@ -31,15 +31,20 @@ namespace NzbDrone.Core.MediaCover
                 return false;
             }
 
+            var headers = _httpClient.Head(new HttpRequest(url)).Headers;
+            var fileSize = _diskProvider.GetFileSize(path);
+            if (fileSize != headers.ContentLength)
+            {
+                return false;
+            }
+            
             if (!IsValidGDIPlusImage(path))
             {
                 _diskProvider.DeleteFile(path);
                 return false;
             }
-
-            var headers = _httpClient.Head(new HttpRequest(url)).Headers;
-            var fileSize = _diskProvider.GetFileSize(path);
-            return fileSize == headers.ContentLength;
+            
+            return true;
         }
 
         private bool IsValidGDIPlusImage(string filename)
